pyproject.toml
[project]
name = "fastapi-rss"
version = "0.1.0"
description = "RSS Reader"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "celery>=5.4.0",
    "feedparser>=6.0.11",
    "sqlalchemy>=2.0.36",
    "duckdb>=1.1.0",
    "redis>=5.2.0",
    "pydantic>=2.9.0",
    "httpx>=0.27.2",
    "python-multipart>=0.0.9",
    "sqladmin>=0.19.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "itsdangerous>=2.1.2",
    "loguru>=0.7.2",
]

[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple/"

[tool.uv.pip]
index-url = "https://mirrors.aliyun.com/pypi/simple/"
